Expose whether limit is met on limiter
This change allows the limiter to be used in a health check
handler or other supervising code to determine if a limit has
been met or not.

Previously the API was encapsulated with a HTTP handler and
private.

Tested with a unit test and end to end with the of-watchdog in
a temporary patch.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>